refactor(data): extract AI-facing resolution/normalization to data/server/resolve (P0)
Verbatim extraction of boundTableIdFor / resolveJurisdiction /
resolveDatabaseRef / findColumn / translateOptionValue / normalizeCellInput /
writeBlockReason from ai/tools/data-tools.ts into
lib/domain/data/server/resolve.ts, so the standalone data tools and the
P1/P2 iteration capture path share one implementation. Refusal strings and
semantics unchanged (they are model-facing contract). data-tools call sites
untouched — ToolExecuteContext satisfies the new DataToolContext
structurally. The plan's 'association-on-approval helper' resolves to
reusing the existing addAutoAssociation (lib/features/conversations) — no
new code needed.

Gates: typecheck clean; lint 0 errors / 151 warnings (ratchet 175, none new).
